+++ config/importmap.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+pin 'country_state_select', to: 'country_state_select/country_state_select.js'
+pin 'country_state_select/controller', to: 'country_state_select/controller.js'