+++ spec/dummy/app/javascript/controllers/country_state_select_controller.js
@@ -1,0 +1,1 @@
+export { default } from "country_state_select/controller";