+++ spec/dummy/db/migrate/20260703180525_create_businesses.rb
@@ -1,0 +1,12 @@
+class CreateBusinesses < ActiveRecord::Migration[7.0]
+  def change
+    create_table :businesses do |t|
+      t.string :name
+      t.string :country
+      t.string :state
+      t.string :city
+
+      t.timestamps
+    end
+  end
+end